(#249) Make size factor of sprite_font_render_text float

+++ src/game/camera.c
@@ -174,7 +174,7 @@ int camera_fill_triangle(camera_t *camera,
 
 int camera_render_text(camera_t *camera,
                        const char *text,
-                       int size,
+                       float size,
                        color_t color,
                        vec_t position)
 {
